feat: Update dependencies and streamline packaging
Updates the `kotlin-stdlib-jdk8` dependency from version `1.8.0` to `2.2.20` across the following modules:
*   `module-b`
*   `module-d`
*   `module-e`
*   `module-f`

In the `app` module, the packaging options are refined by adding an exclusion for `META-INF/INDEX.LIST` and removing a `JavaCompile` configuration that set the `-Xlint:-deprecation` compiler argument.

Additionally, the `logback-classic` runtime dependency is removed from the `data/api` module.